reorganised files into “app” structure
