+++ Dockerfile
@@ -1,5 +1,5 @@
 #======================== builder ========================
-FROM python:3.11.2-bullseye as builder
+FROM python:3.11.3-bullseye as builder
 
 RUN python -m venv --copies /venv
 RUN /venv/bin/pip install --upgrade pip
@@ -17,11 +17,11 @@ RUN /venv/bin/poetry config virtualenvs.create false \
 
 #======================== multistage ========================
 #FROM cgr.dev/chainguard/python
-FROM python:3.11.2-slim-bullseye
+FROM python:3.11.3-slim-bullseye
 COPY --from=builder-venv /venv /venv
 
 COPY . /app
 WORKDIR /app
 
 # And then will start Gunicorn with Uvicorn
-ENTRYPOINT ["/venv/bin/gunicorn"  , "-k", "uvicorn.workers.UvicornWorker", "-c", "docker-images/gunicorn_conf.py", "main:app"]
+ENTRYPOINT ["/venv/bin/gunicorn"  , "-k", "uvicorn.workers.UvicornWorker", "-c", "app/docker-images/gunicorn_conf.py", "main:app"]
